Updated openstack/openstack Project: openstack/neutron-lbaas  c7846671a77b9498a58c09a1b3543b1f72946fb2
Agent, device driver interface, and callback API

LBaaS V2 object model needed an agent to redone to work with
the new object model.  This creates the agent main script with
a place holder agent manager used as the service.  This also
adds the interface for device drivers.  The first device driver
implementation will the be namespace haproxy driver redone for
the new object model but will be added in a later review.  Lastly,
this adds the API class that exposes the plugin callbacks.

Partially-implements: blueprint lbaas-refactor-haproxy-namespace-driver-to-new-driver-interface
Co-Authored-By: Phillip Toohill <phillip.toohill@rackspace.com>
Co-Authored-By: Brandon Logan <brandon.logan@rackspace.com>
Change-Id: Ie7842bd347380a7c7ed1c32c9b63ceb0adaaa58f