Updated openstack/openstack Project: openstack/neutron-lbaas  03307701d480bf2e5fee6f0b1291abda9924694a
Completion of Agent Manager

The agent manager is responsible for receiving and action on messages
from the queue.  Most of the time the action is delegating to the
device driver to do the action.  The agent manager is responsible
for the final status management.  Whether an action succeeds or
fails, the agent manager should be responsible for setting the
appropriate status.  It will also periodically check if
load balancers are out of sync and periodically grab stats and
health information of members.

Change-Id: I0fbd2914381b49ffbc651a0684ce341b4a651f2e
Partially-implements: blueprint lbaas-refactor-haproxy-namespace-driver-to-new-driver-interface
Co-Authored-By: Brandon Logan <brandon.logan@rackspace.com>