Updated openstack/openstack Project: openstack/heat  6950453e1282a8f6992651c814e14fbdee3f3590
Utility function for int param checking

This patch adds a utility function for checking the int type parameters
passed to the API layer. A test case is included for this utility
function. To exemplify how to use it, the events controller is revised
to make use of this function. The plan is to add this checking to all
int type parameters at API layer.

Change-Id: If4a1e2e5e7adbc272e2cfa5b1918cdf733926013