Updated openstack/openstack Project: openstack/barbican  90b4cc002362b182deaf5de25beaaeaf874ca4d3
Fixed Binary Encoding to Secret Stores

The SecretDTO should encapsulate the secret as a Base64 string.  The
translations module was decoding Base64 secrets before inserting them
into the SecretDTO that was later passed to the SecretStores, and it
was not encoding plain text into Base64 as well. This code fixes that.

Closes-Bug: #1423941
Change-Id: I3f2f86a0bf93b44b12b0c23f53fb617d9b73043f
Co-authored-by: Kaitlin Farr <kaitlin.farr@jhuapl.edu>
Related: blueprint content-types