Updated openstack/openstack Project: openstack/tripleo-image-elements  5c1f414e9da3096aa0492b698d0e3ea6414503d9
Drop augtools from the tempest element

Since passthrough for tempest has been introduced, it is possible
for a section and/or a setting to appear more than once in
tempest.conf, which augtool cannot handle.

In fact it is not necessary to inject configuration settings, it
is fine to just append them to the config file, as the last
setting wins in any case.

Change-Id: Ie7bdb1d327c71a6b32d0b947d0f1764f9b14b40c