Updated openstack/openstack Project: openstack/cinder  399bf5b1b4f7b0926b48fa0426aeb4e4ab149a89
Dell SC API change fails snapshot creation

The Dell Storage Center driver is failing snapshot creation on
later versions of the Dell REST API.

The doNotExpire option was removed in API 15.3.1.110.  We were
setting this so that the replays are not expired and removed when
Cinder expects them to still be there.  This has been replaced by
setting expiretime to 0.  We were doing this already so the fix
was to simply remove doNotExpire from the REST request payload.

This will fail on every future release of the Dell Storage Center
so this is a critical fix.

Closes-Bug 1428259

Change-Id: I51fbf599e29e30793cdd1e6014e876d749732fd3