Updated openstack/openstack Project: openstack/glance  2420b6b32906a5fd434862ef0dfed3364b57f0cc
The metadef tags create api does not match blue-print

The blueprint for metadef tags shows:
Create / Replace all tags in a specific namespace
    POST /metadefs/namespaces/{namespace}/tags/
and
Add tag in a specific namespace:
    POST /metadefs/namespaces/{namespace}/tags/{tag}

Unfortunately, these were implemented as:
Create / Replace all tags in a specific namespace
    PUT /metadefs/namespaces/{namespace}/tags/
and
    POST /metadefs/namespaces/{namespace}/tags/

This bug seeks to rectify this situation and use the correct
action/URI. Even though the metadef tag library has been checked in,
the associated python-glanceclient check in has not been approved yet.
Additionally, the specs for the metadef tag changes have not been posted
anywhere, so, hopefully this can be changed with minimal impact.

Change-Id: I7d46c529e50b69a386ad7162c65187cec4b66249
Closes-Bug: 1419945