Updated openstack/openstack Project: openstack/ironic  dbd53ca9f87abd0a7e2fffe6a58a908f59207cf3
Add tftp mapfile configuration in install-guide

The document http://docs.openstack.org/developer/ironic/deploy/install-guide.html
just show how to create a tftp map file, but this have to be enabled in tftp
configuration file, else, it is not working as well.

In this patch, add the tftp map file configuration steps.
Closes-Bug: #1427953

Change-Id: I3d838580cbd60b3264b9378a547cadbae2136d36