Updated openstack/openstack Project: openstack/python-troveclient  88a516dc62e1f5d2280351dcb9abdb18fa5ec953
Clean up remaining references to TROVE_URL

Replaced last remaining reference to TROVE_URL in python-troveclient.
Currently the user along with OS_URL , or TROVE_URL , is used as part
of a hash for a unique directory for the bash completion cache. Since
I'm not sure if there is a reason why the legacy URLs are used here,
I've replaced TROVE_URL with NOVA_URL instead of OS_AUTH_URL which
might be more appropriate for this purpose.

Change-Id: Id30aeae992aecba3c1fbe53c2ce421157410ca20
Closes-Bug: #1238282