Updated openstack/openstack Project: openstack-dev/devstack  7a678bfcb2726506a6daff3893c962ade4abb96c
Do not install the python-virtualenv on Fedora

The files/rpms/general:python-virtualenv explicitly
installed this package on Fedoras.

This package is not installed on other distros by devstack
by default.

If you stack/unstack the package gets reinstalled,
and confuses the system about the installed virtual-env version.
The uninstall works in CI, but it can be problematic when you do
reinstalls on the same machine.

The uninstall introduced by
834b804d3eda9029d3c66db0ab732a76a22ed08b, this commit deos not
has any reference to the external bug what it supposed to solve.

Related RDO thread started here:
https://www.redhat.com/archives/rdo-list/2015-March/msg00015.html

Change-Id: I4a723f179bdc28d39a4910fb9e3787e9e67c354b