Updated openstack/openstack Project: openstack/barbican  3ce9b8060546ee9531afb13a17e51222be19f158
Ported API documentation to the repo

The contents from the API documentation in the CloudKeep repo[1] were
copied as they are at the moment and translated into a format that's
usable for Sphinx. The intent was not to have the updated API, but to
have the documentation in this repository. Further commits will be
made updating the contents.

[1] https://github.com/cloudkeep/barbican/wiki/Application-Programming-Interface

Change-Id: I42747081348b1d02c31b368302837ea7b0b54467