Updated openstack/openstack Project: openstack/neutron  56967bd950b5de37a244333349d4969ec75be2a9
Fix test tautology for DVR

Change 17cae2c added tests that pass regardless of whatever code they
are supposed to validate. If one asked for a bad example of how not to
write/review test, change 17cae2c would be one of it.

This fix is my way to repent.

Change-Id: Ibb27f24380beab99388c0a6dc187b75e0852761e