Updated openstack/openstack Project: openstack/tripleo-incubator  310b557430a7dc6fbf6db892fde7fd4ad7ce6a58
Fall back to generated snmp password with --no-undercloud

Running devtest_overcloud.sh via devtest with --no-undercloud seems
to result in some puppet errors recently, due to the zero-length
password in the heiradata key.

So instead of passing a zero length password, fall back to generating
one in the event no password is found via the undercloud key of the
test environment.

Change-Id: Ia5f9fff1ea3b50f372489a42647e3f47bc219df8
Closes-Bug: #1428726