Updated openstack/openstack Project: openstack/neutron  0a2a0aca61cd32aba541812d98514cfc1f06580a
Add InvalidIpForNetwork and InvalidIpForSubnet exception

When posting an 'attach interface' request to Nova with an invalid
ip for defined network, Nova returns an HTTP 500 error. In fact,
Neutron returns 'InvalidInput' error, but Neutron client is not able
to translate this error to a specific exception. So that a general
'BadRequest' exception is thrown. Neutron client and Nova need a more
specific Neutron error type, in order to address and translate the
error in to a proper Nova exception.

APIImpact

Change-Id: Id4d3108234a198c0eb75237ea8991a72bd3cc4bb
Partial-Bug: 1369871