Updated openstack/openstack Project: openstack/oslo.utils  91dc782c2f11fb76425effd7a331d63111adf1ce
Add a stopwatch + split for duration(s)

The existing timeutils provides a way to measure durations
that is also used to get actual times and dates; for most
cases this dual ability is not needed and the user only cares
about the duration between two time points. In cases where
ntpd may affect the machine clock this usage is also incorrect
and may go backwards in time...

To support this kind of specific and frequently occurring usage
a stopwatch class can be used that can enable more accurate
gathering of *just* durations (and using monotonic time when
it's available in python 3.3+ or optionally when the library
at http://pypi.python.org/pypi/monotonic is installed).

This code comes from taskflow, but oslo.messaging has similar
code that they could replace with this (and I imagine tooz
could also use this); among others...

Change-Id: I950e26c1eb8283a134fb4f16544f87bc5d29fd82