Updated openstack/openstack Project: openstack/nova  f480545df79bb56f5f6bf5973ec5e7f658a42b19
Allow disabling the evacuate cleanup mechanism in compute manager

This mechanism attempts to destroy any locally-running instances on
startup if instance.host != self.host. The assumption is that the
instance has been evacuated and is safely running elsewhere. This is
a dangerous assumption to make, so this patch adds a configuration
variable to disable this behavior if it's not desired.

Note that disabling it may have implications for the case where
instances *were* evacuated, given potential shared resources.
To counter that problem, this patch also makes _init_instance()
skip initialization of the instance if it appears to be owned
by another host, logging a prominent warning in that case.

As a result, if you have destroy_after_evacuate=False and you start
a nova compute with an incorrect hostname, or run it twice from
another host, then the worst that will happen is you get log
warnings about the instances on the host being ignored. This should
be an indication that something is wrong, but still allow for
fixing it without any loss. If the configuration option is disabled
and a legitimate evacuation does occur, simply enabling it and then
restarting the compute service will cause the cleanup to occur.

This is added to the workarounds config group because it is really
only relevant while evacuate is fundamentally broken in this way.
It needs to be refactored to be more robust, and once that is done,
this should be able to go away.

DocImpact: New configuration option, and peril warning
Partial-Bug: #1419785
Change-Id: Ib9a3c72c096822dd5c65c905117ae14994c73e99