Updated openstack/openstack Project: openstack/trove  37e0980d6ec7eafeb6634e5e86390ba5304eeac7
Introduce a classification of datastores and strategies

This change introduces a classification of datastores and strategies
based on our understanding of the readiness of that particular item of
code. It reflects the conversation at the mid-cycle in Seattle (Kilo).

A specification for this change was submitted for review in
https://review.openstack.org/#/c/154119/

Implements: blueprint experimental-datastores
DocImpact
Change-Id: Id710bdd21070c3af160f1594d29c5ac41d5c3548