Updated openstack/openstack Project: openstack/python-ironicclient  d76862b1869ea529da1e149f4e3a00cc5fb9eda6
Adds basic ironicclient functional testing

This moves unit tests to ironicclient/tests/unit and adds a new test
directory at ironicclient/tests/functional.  These are meant to be
functional tests that are run against a living cloud (presumably deployed
by devstack as part of a gating job).  They depend on a test.conf to exist
that contains admin user credentials for said cloud.  This includes a simple
run_functional.sh script that is meant to create that and serve as the entry
point for running the tests from a devstack post_test_hook.

Change-Id: I11570c3e22e4f80b94d0643dca6ed8231217c877