Updated openstack/openstack Project: openstack/heat  7f0ff062479a1bce536aac3b2684ae9424ed3bac
Add extra data to Stack table for Convergence.

The convergence design requires extra data to be stored with each Stack
row in order to manage concurrent updates.

In the update query for updating the stack, session synchronization
is set to False to avoid error causing due to status_reason column
inherited in Stack table from the StateAware table. Updates to
status_reason column is handled via @property and this causes problems
due to the fact that the UPDATE...WHERE updates it in DB but fails to
correctly evaluate the sqlalchemy object directly. The evaluation fails
with "KeyError" on status_reason column. By setting synchronization off
and by expiring the session, we enforce status_reason to be read from DB
again.

Implements: blueprint convergence-stack-data

Change-Id: I5f7482b94d76d937f03772596873c46bd536ac4f