Updated openstack/openstack Project: openstack/nova  281f6cdd5cec1c22c775a633a32b36f087e5bd51
Add API Response class for more complex testing

the api client .api_(get|delete|post|put) do useful encoding/decoding
of the request/response from dictionaries, and to dictionaries. For
negative and functional testing we're going to often need the status
code as well.

This creates an APIResponse class which wraps and decodes the Requests
library response that is returned. That provides easy access to the
decoded json, as well as access to things like status code and
headers.

It's used in follow on tests to make them more concise.

The convenience methods in the test client (such as get_server) are
adapted to return the same information as before, but through this new
method. These will be further documented in future commits.

Change-Id: I692cbc1a39c3802799de45b4c40306fd717feabf