Updated openstack/openstack Project: openstack/barbican  e250ff6c040dca2ca875920cb2bac37a8cd9cfba
Ensure that external secret refs cannot be added to containers

Prior to this change, it was possible for one to point secret_refs to
arbitrary locations, especially in the context of adding secrets to a
container given that the hostname part of the secret_ref was being
ignored. Since we are now checking secret_refs against the
configured hostname of the Barbican instance, this is no longer
possible. In other words, this is just some extra validation to ensure
that parameters remain consistent with the Barbican instance
configuration.
Closes-Bug: 1423282

Change-Id: Ife1d66f234050e3150a879e76142b68f8f72a524