Updated openstack/openstack Project: openstack/neutron  750910c6808016a43c15f8107da97163ed511f08
Decouple L3 and VPN service plugins during router operations

This change leverages the event registry to decouple L3 and VPN
when checking whether certain router operations can be permitted
or not. Follow-up changes will clean up the rest of the coupling.

This clean-up is required because the existing code couples the two
service plugins together, as L3 needs to know about a higher-level
service, like VPN. A similar coupling would be needed for any other
service that might be using or interested in knowning about router
resources; in a world where advanced services have a life of their
own this hardly works or it leads to the tight coupling
experienced so far, which is not the right way to go about it.

Depends-on: I6ac366b329466df6c79a3683e6374feca76773a0
Related-blueprint: services-split
Related-blueprint: plugin-interface-perestroika

Change-Id: Iada03606982670111ae911549a3aad85ed358bea