Updated openstack/openstack Project: openstack/tempest  15494ac672ad7a60e8165e74aabfab210b4b7b2d
Check Floating ip status before load balancer check

This patch checks for the floating ip status before
the load balancer check.
This would make sure that the floating ip is in an
active state before checking the connection to the
load balancer vip.

If there is an issue with the Floating IP that would
be more evident from the error message.

This would also give enough time for the floatingip
to become active.

Partial-Bug: #1426764
Change-Id: Ic25d51b6d448b156df14d3eaf2e8c32e3ca61a07