Updated openstack/openstack Project: openstack/ironic  5b8e9bd5c2fad57278c5aea977c4663422deb2aa
Add server-supported min and max API version to HTTPNotAcceptable(406)

Add server-supported min and max API version headers to the response
associated with HTTPNotAcceptable exceptions. These exceptions could
happen when the request specifies an unsupported API version.
The returned headers will provide feedback to the caller as to
what the acceptable versions are.

Change-Id: I84e35af2c3e97ba70bb91b6ce32f13ca307201c5