Updated openstack/openstack Project: openstack/nova  60ef65b6abbc7114f55e16d9205456c0492e460a
Compute: update config drive settings on instance

In the case the config drive setting are overridden by either
image properties or the configuration setting then these are updated
on the instance to actually reflect the running VM configuration.

The patch creates a common method for updating instance details
after spawn.

Change-Id: Ia5e36989a7b7543fa77c7c196919f8566e02d28a
Closes-bug: #1403889