Updated openstack/openstack Project: openstack/nova  447010a4f76ffe66a98efd2973e1a91181767baf
xenapi: Fix session tests leaking state

The current tests are setting the `errno` on the `socket.error` *class* which
leaks across tests. This causes the `test_virt_driver.py` tests to fail if run
after the `test_session` tests (which the gate does not do).

The solution is to set the `errno` attribute test on the *instance* so that no
leakage occurs.

Change-Id: I9570523e717db3d436649c0ec396ad1b8152a1c4