Updated openstack/openstack Project: openstack/oslo.messaging  883318bd9606740576bfa1eeb931b635166adf37
Use import of zmq package for test skip

Using the import of impl_zmq to decide whether to skip tests is
unsafe as if something breaks in the import, then the tests are
skipped, potentially hiding a bug.

Switch to using the import of the zmq package itself to decide
whether tests should be executed or not; this should cover the
currently unsupported Python3 test coverage.

Change-Id: I7d65506845b039f5e2be4c460bede977122deca9