Updated openstack/openstack Project: openstack/nova  fccb9ef3f155a11c0d8218008b55c7c25f4a61f4
Change instance disappeared during destroy from Warning to Info

Commit 2fee082e9267c75a1e676361d975e52ac9ad86f8 changed this log message
from error to warning because it's an expected situation while powering
off an instance.

In 7 days, the warning shows up over 122 thousand times in the gate with
91% of those runs being successful, so it's extremely doubtful that this
situation is a contributing factor to race bugs in the gate.

Drop the log level from warning to info since it's useful to have
something in the logs to indicate that the instance is gone as a unit of
work.

Change-Id: I82560c9f2b2efd619e3565f517d6d2936f1be206