Updated openstack/openstack Project: openstack/keystone  d573ce7b3193dd3e3ad562b977ef9f73f854926f
Fixes bug in SQL/LDAP when honoring driver_hints

SQL and LDAP backends failed to honor multiple
valued driver_hints due to a wrong manipulation
of the list of filters.

When iterating over the list of filters, the
method removes elements from it at the same time.
As result, filters in odd indexes are skipped.

This patch fixes this bug and assert the expected
behavior.

Closes-Bug: #1424745

Change-Id: Idc755de01d5dfd1013d6825c2c77436b256da096