Updated openstack/openstack Project: openstack-infra/puppet-subunit2sql  d286d9f63da8d56ede54e80ebbd81309a64d9911
Add option to disable uneccesary data migration

The latest subunit2sql migration includes a data migration to extract
the microseconds from timestamps into a separate column. However, the
infra DB for this has already stripped out all the microsecond data
from the timestamps, making running that part of the migration
unnecessary. This commit adds the flag to the template subunit2sql
config file so that we don't run that and just use the default value,
0, for the new columns, which should save a significant amount of time
when running the migration.

Change-Id: Idf32b883babc79cabe3da1d38e3f7ddd5c4127de