Updated openstack/openstack Project: openstack/neutron  fbd442408aca567c5e64a73c971382a8b84d421a
oslo: sync all modules that depend on incubator log module

We're going to switch to oslo.log, so we need to untangle incubator
modules from the old log module. This is achieved by synchronizing all
the modules that used the module before from oslo-incubator.

Note that incubator modules now use standard logging module (and not
e.g. oslo.log).

The following modules (and their dependencies) are synced:
- eventlet_backdoor
- fileutils
- loopingcall
- periodic_task
- policy
- service
- systemd
- threadgroup
- versionutils

The latest commit in oslo-incubator at the moment of sync is:
- f989c4397d7e83c3e73e8da01a9f27bf4ca38b73

Changes to modules are minor and hence are not expected to break
anything.

Change-Id: I360863a5a4fa6a179c6884cff84553a35d889719
Related-Bug: #1425013